Move ardin_mb and stone_cx into projects/
