+++ projects/ardin_mb/perfect_memory.cc
