+++ projects/ardin_mb/plot_active_neurons.py
