+++ projects/ardin_mb/plot_terminal_state.py
