+++ projects/ardin_mb/sim_params.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // Third-party includes
-#include "../third_party/units.h"
+#include "third_party/units.h"
 
 //------------------------------------------------------------------------
 // SimParams
