+++ projects/stone_cx/MakefileReplay
@@ -1,7 +1,7 @@
-EXECUTABLE      := simulatorVicon
-SOURCES         := simulatorVicon.cc simulatorCommon.cc robotCommon.cc
+EXECUTABLE      := simulatorReplay
+SOURCES         := simulatorReplay.cc simulatorCommon.cc visualizationCommon.cc
 LINK_FLAGS      := `pkg-config --libs opencv`
-CXXFLAGS        := `pkg-config --cflags opencv` -pthread
+CXXFLAGS        := `pkg-config --cflags opencv` -I../..
 CPU_ONLY        =1
 CPP_STANDARD    =c++14
