+++ projects/stone_cx/MakefileRobot
@@ -1,7 +1,7 @@
-EXECUTABLE      := simulatorRobotDeadReckon
-SOURCES         := simulatorRobotDeadReckon.cc simulatorCommon.cc robotCommon.cc visualizationCommon.cc
+EXECUTABLE      := simulatorRobot
+SOURCES         := simulatorRobot.cc simulatorCommon.cc robotCommon.cc
 LINK_FLAGS      := `pkg-config --libs opencv`
-CXXFLAGS        := -pthread `pkg-config --cflags opencv`
+CXXFLAGS        := `pkg-config --cflags opencv` -pthread -I../..
 CPU_ONLY        =1
 CPP_STANDARD    =c++14
 
@@ -11,6 +11,10 @@ endif
 
 ifdef RECORD_SENSORS
     CXXFLAGS += -DRECORD_SENSORS
+endif
+
+ifdef USE_SEE3_CAM
+    CXXFLAGS += -DUSE_SEE3_CAM
 endif
 
 include $(GENN_PATH)/userproject/include/makefile_common_gnu.mk
