+++ projects/stone_cx/MakefileVicon
@@ -1,7 +1,7 @@
-EXECUTABLE      := simulatorReplay
-SOURCES         := simulatorReplay.cc simulatorCommon.cc visualizationCommon.cc
+EXECUTABLE      := simulatorVicon
+SOURCES         := simulatorVicon.cc simulatorCommon.cc robotCommon.cc
 LINK_FLAGS      := `pkg-config --libs opencv`
-CXXFLAGS        := `pkg-config --cflags opencv`
+CXXFLAGS        := `pkg-config --cflags opencv` -pthread -I../..
 CPU_ONLY        =1
 CPP_STANDARD    =c++14
