+++ projects/stone_cx/plot_electrophys.py
