+++ projects/stone_cx/plot_vicon.py
