+++ projects/stone_cx/simulator.cc
@@ -1,9 +1,7 @@
 // Standard C++ includes
-#include <fstream>
 #include <iostream>
 #include <numeric>
 #include <random>
-#include <sstream>
 #include <vector>
 
 // Standard C includes
@@ -14,7 +12,8 @@
 #include <opencv2/opencv.hpp>
 
 // Common includes
-#include "../genn_utils/analogue_csv_recorder.h"
+#include "common/von_mises_distribution.h"
+#include "genn_utils/analogue_csv_recorder.h"
 
 // GeNN generated code includes
 #include "stone_cx_CODE/definitions.h"
@@ -22,20 +21,11 @@
 // Model includes
 #include "parameters.h"
 #include "simulatorCommon.h"
+#include "spline.h"
 #include "visualizationCommon.h"
 
 using namespace BoBRobotics;
 using namespace BoBRobotics::StoneCX;
-
-namespace
-{
-double readDoubleField(std::istringstream &lineStream)
-{
-    std::string field;
-    std::getline(lineStream, field, ',');
-    return std::stod(field);
-}
-}
 
 int main()
 {
@@ -46,8 +36,17 @@ int main()
     
     const double preferredAngleTN2[] = { Parameters::pi / 4.0, -Parameters::pi / 4.0 };
     
+    // Outbound path generation parameters
+    const unsigned int numOutwardTimesteps = 1500;
+
+    // Agent dynamics parameters
+    const double pathLambda = 0.4;
+    const double pathKappa = 100.0;
+
     const double agentDrag = 0.15;
 
+    const double agentMinAcceleration = 0.0;
+    const double agentMaxAcceleration = 0.15;
     const double agentM = 0.5;
     
     allocateMem();
@@ -77,29 +76,58 @@ int main()
     cv::moveWindow("Activity", pathImageSize, 0);
     cv::Mat activityImage(activityImageHeight, activityImageWidth, CV_8UC3, cv::Scalar::all(0));
 
+    // Create Von Mises distribution to sample angular acceleration from
+    std::array<uint32_t, std::mt19937::state_size> seedData;
+    std::random_device seedSource;
+    std::generate(seedData.begin(), seedData.end(),
+                  [&seedSource](){ return seedSource(); });
+    std::seed_seq seeds(std::begin(seedData), std::end(seedData));
+    std::mt19937 gen(seeds);
+
+    VonMisesDistribution<double> pathVonMises(0.0, pathKappa);
+
+    // Create acceleration spline
+    tk::spline accelerationSpline;
+    {
+        // Create vectors to hold the times at which linear acceleration
+        // should change and it's values at those time
+        const unsigned int numAccelerationChanges = numOutwardTimesteps / 50;
+        std::vector<double> accelerationTime(numAccelerationChanges);
+        std::vector<double> accelerationMagnitude(numAccelerationChanges);
+
+        // Draw accelerations from real distribution
+        std::uniform_real_distribution<double> acceleration(agentMinAcceleration,
+                                                            agentMaxAcceleration);
+        std::generate(accelerationMagnitude.begin(), accelerationMagnitude.end(),
+                      [&gen, &acceleration](){ return acceleration(gen); });
+
+        for(unsigned int i = 0; i < numAccelerationChanges; i++) {
+            accelerationTime[i] = i * 50;
+        }
+
+        // Build spline from these
+        accelerationSpline.set_points(accelerationTime, accelerationMagnitude);
+    }
+
 #ifdef RECORD_ELECTROPHYS
-    GeNNUtils::AnalogueCSVRecorder<scalar> tn2Recorder("tn2.csv", rTN2, Parameters::numTN2, "TN2");
-    GeNNUtils::AnalogueCSVRecorder<scalar> cl1Recorder("cl1.csv", rCL1, Parameters::numCL1, "CL1");
-    GeNNUtils::AnalogueCSVRecorder<scalar> tb1Recorder("tb1.csv", rTB1, Parameters::numTB1, "TB1");
-    GeNNUtils::AnalogueCSVRecorder<scalar> cpu4Recorder("cpu4.csv", rCPU4, Parameters::numCPU4, "CPU4");
-    GeNNUtils::AnalogueCSVRecorder<scalar> cpu1Recorder("cpu1.csv", rCPU1, Parameters::numCPU1, "CPU1");
+    AnalogueCSVRecorder<scalar> tn2Recorder("tn2.csv", rTN2, Parameters::numTN2, "TN2");
+    AnalogueCSVRecorder<scalar> cl1Recorder("cl1.csv", rCL1, Parameters::numCL1, "CL1");
+    AnalogueCSVRecorder<scalar> tb1Recorder("tb1.csv", rTB1, Parameters::numTB1, "TB1");
+    AnalogueCSVRecorder<scalar> cpu4Recorder("cpu4.csv", rCPU4, Parameters::numCPU4, "CPU4");
+    AnalogueCSVRecorder<scalar> cpu1Recorder("cpu1.csv", rCPU1, Parameters::numCPU1, "CPU1");
 #endif  // RECORD_ELECTROPHYS
 
-    std::ifstream replayData("data.csv");
-
     // Simulate
+    double omega = 0.0;
     double theta = 0.0;
     double xVelocity = 0.0;
     double yVelocity = 0.0;
     double xPosition = 0.0;
     double yPosition = 0.0;
-    double xStartPosition = 0.0;
-    double yStartPosition = 0.0;
-    bool first = true;
     for(unsigned int i = 0;; i++) {
         // Project velocity onto each TN2 cell's preferred angle and use as speed input
         for(unsigned int j = 0; j < Parameters::numTN2; j++) {
-            speedTN2[j] = (sin(theta + preferredAngleTN2[j]) * xVelocity) +
+            speedTN2[j] = (sin(theta + preferredAngleTN2[j]) * xVelocity) + 
                 (cos(theta + preferredAngleTN2[j]) * yVelocity);
         }
 
@@ -120,40 +148,17 @@ int main()
         // Visualize network activity
         visualize(activityImage);
 
-        // If we there is more data to replay
-        std::string line;
-        const bool outbound = !std::getline(replayData, line).fail();
+        // If we are on outbound segment of route
+        const bool outbound = (i < numOutwardTimesteps);
+        double a = 0.0;
         if(outbound) {
-            // Read position from file
-            std::istringstream lineStream(line);
-            xPosition = readDoubleField(lineStream);
-            yPosition = readDoubleField(lineStream);
-            readDoubleField(lineStream);
+            // Update angular velocity
+            omega = (pathLambda * omega) + pathVonMises(gen);
 
-            // If this is the first position, update start position so drawing can be centred
-            if(first) {
-                xStartPosition = xPosition;
-                yStartPosition = yPosition;
-                first = false;
-            }
+            // Read linear acceleration off spline
+            a = accelerationSpline((double)i);
 
-            // Read velocity
-            xVelocity = 6.0f * readDoubleField(lineStream);
-            yVelocity = 6.0f * readDoubleField(lineStream);
-            readDoubleField(lineStream);
-
-            // Read orientation, flipping and converting from (-pi,pi) to (0, 2pi)
-            readDoubleField(lineStream);
-            readDoubleField(lineStream);
-            theta = -readDoubleField(lineStream);
-            if(theta < 0.0) {
-                theta = (2.0 * Parameters::pi) + theta;
-            }
-
-            // If this is the end of outbound data, zero velocity (as simulated dynamics don't match well) and display CPU4 high-water mark
-            if(replayData.eof()) {
-                xVelocity = 0.0;
-                yVelocity = 0.0;
+            if(i == (numOutwardTimesteps - 1)) {
                 std::cout << "Max CPU4 level r=" << *std::max_element(&rCPU4[0], &rCPU4[Parameters::numCPU4]) << ", i=" << *std::max_element(&iCPU4[0], &iCPU4[Parameters::numCPU4]) << std::endl;
             }
         }
@@ -164,28 +169,28 @@ int main()
             const scalar rightMotor = std::accumulate(&rCPU1[8], &rCPU1[16], 0.0f);
 
             // Use difference between left and right to calculate angular velocity
-            const double omega = -agentM * (rightMotor - leftMotor);
+            omega = -agentM * (rightMotor - leftMotor);
 
             // Use fixed acceleration
-            const double a = 0.1;
-
-            // Update heading
-            theta += omega;
-
-            // Update linear velocity
-            // **NOTE** this comes from https://github.com/InsectRobotics/path-integration/blob/master/bee_simulator.py#L77-L83 rather than the methods section
-            xVelocity += sin(theta) * a;
-            yVelocity += cos(theta) * a;
-            xVelocity -= agentDrag * xVelocity;
-            yVelocity -= agentDrag * yVelocity;
-
-            // Update position
-            xPosition += xVelocity;
-            yPosition += yVelocity;
+            a = 0.1;
         }
 
+        // Update heading
+        theta += omega;
+
+        // Update linear velocity
+        // **NOTE** this comes from https://github.com/InsectRobotics/path-integration/blob/master/bee_simulator.py#L77-L83 rather than the methods section
+        xVelocity += sin(theta) * a;
+        yVelocity += cos(theta) * a;
+        xVelocity -= agentDrag * xVelocity;
+        yVelocity -= agentDrag * yVelocity;
+
+        // Update position
+        xPosition += xVelocity;
+        yPosition += yVelocity;
+
         // Draw agent position (centring so origin is in centre of path image)
-        const cv::Point p((pathImageSize / 2) + (int)(0.25 * (xPosition - xStartPosition)), (pathImageSize / 2) + (int)(0.25 * (yPosition - yStartPosition)));
+        const cv::Point p((pathImageSize / 2) + (int)xPosition, (pathImageSize / 2) + (int)yPosition);
         cv::line(pathImage, p, p,
                  outbound ? CV_RGB(0xFF, 0, 0) : CV_RGB(0, 0xFF, 0));
