+++ projects/stone_cx/visualizationCommon.cc
